+++ index.ts
@@ -22,10 +22,10 @@ export { infoMessage } from "./src/cli/ui/consoleOutput";
 export { sdmCd } from "./src/pack/sdm-cd/support/SdmCd";
 export { LocalSdmConfig } from "./src/sdm/configuration/localSdmConfig";
 export {
-    YargSaver,
-    freshYargSaver,
+    YargBuilder,
+    freshYargBuilder,
     yargCommandFromSentence,
     yargCommandWithPositionalArguments,
     ParameterOptions,
     CommandLineParameter,
-} from "./src/cli/invocation/command/support/yargSaver";
+} from "./src/cli/invocation/command/support/yargBuilder";
