--- src/cli/invocation/command/support/yargSaver/index.ts
@@ -1,9 +1,0 @@
-export {
-    YargSaver,
-    freshYargSaver,
-    yargCommandFromSentence,
-    yargCommandWithPositionalArguments,
-    CommandLineParameter,
-    ParameterOptions,
-    isYargSaver
-} from "./YargSaver"
